Combine list tiles, begin sanitize
